+++ handle_string.c
@@ -26,13 +26,13 @@ char	*ft_s(t_struct *form, char *ch)
 	f = ft_strlen(b);
 	if (form->min > form->p && form->min != 0 && form->p > 0)
 		f = form->min;
-	else if (form->p > form->min && form->p < ft_strlen(b))
+	else if (form->p > form->min && form->p < (int)ft_strlen(b))
 		f = form->p;
-	if (form->p > ft_strlen(b))
+	if (form->p > (int)ft_strlen(b))
 		form->p = 0;
 	s = ft_strnew(f);
 	(form->zero && !form->minus) ? ft_memset(s, '0', f) : ft_memset(s, ' ', f);
-	if (form->p < ft_strlen(b) && form->p > 0)
+	if (form->p < (int)ft_strlen(b) && form->p > 0)
 		ft_memset(b + form->p, '\0', ft_strlen(b) - form->p);
 	(form->minus) ? (my_strcpy(s, b)) : (my_strcpy(s + (f - ft_strlen(b)), b));
 	free(b);
